// 05-heroesApp/src/app/heroes/interfaces/heroes.interfaces.ts
export interface Heroe {
    id?:               string;
    superhero:        string;
    publisher:        Publisher;
    alter_ego:        string;
    first_appearance: string;
    characters: string;
    alt_img?: string; //https://iasdawds/img.png
}

export enum Publisher {
    DCComics = "DC Comics",
    MarvelComics = "Marvel Comics",
}